requirements.txt
pyyaml
numpy
google-cloud-speech
azure-cognitiveservices-speech
boto3